requirements.txt
Django==1.9.7
raphendyr-django-essentials==1.3.0
celery==3.1.23
requests==2.10.0
lxml ~= 4.3.2
PyYAML==3.11
docutils==0.12
python-magic==0.4.11
